rttg

+++ src/Controller/MainController.php
@@ -18,11 +18,100 @@ use Symfony\Component\Routing\Annotation\Route;
 class MainController extends AbstractController
 {
     #[Route('/accueil', name: 'app_main')]
-    public function index(Eleves $eleves, ElevesRepository $elevesRepository): Response
+    public function index( ElevesRepository $elevesRepository, Request $request, EntityManagerInterface $entityManager, PictureService $pictureService): Response
     {
-        return $this->render('main/index.html.twig', [
+        $eleves = new Eleves;
+
+        $form = $this->createForm(ElevesFormType::class, $eleves);
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+
+            // on récupère le fichier envoyer dans le champ 'photo'
+            $image = $form->get('photo')->getData();
+
+            // on vérifie qu'il y ait un fichier envoyé
+            if (!empty($image)) {
+
+                $folder = 'image';
+
+                // on récupère l'ancienne photo
+                $previmage = $eleves->getPhoto();
+
+                // on supprime l'ancienne photo de profil
+                $pictureService->delete($previmage, $folder);
+
+                // on ajoute la nouvelle photo de profil
+                $fichier = $pictureService->add($image, $folder, 300, 300);
+
+                $eleves->setPhoto($fichier);
+
+            }
+
+            // ajouter un transport 
+            $newtransport = $form->get('newtransport')->getData();
+
+            if (!empty($newtransport)) {
+                $nouvelleEntite = new Transports();
+                $nouvelleEntite->setTransport($newtransport);
+                $eleves->addTransport($nouvelleEntite);
+                $entityManager->persist($nouvelleEntite);
+            }
+
+            // enlever un transport
+            $removetransport = $form->get('transports')->getData();
+            
+            if(!empty($removetransport)){
+                foreach($removetransport as $rt){
+                    $rt->setEleves(null);
+                    $entityManager->remove($rt);
+                }
+            }
+
+            // la date de naissance
+            $date = $form->get('date_naissance')->getData();
+            
+            $date = DateTime::createFromFormat("d/m/Y",$date);
+
+            if(!empty($date)){
+                $eleves->setDateNaissance($date);
+            }
+
+            // la date d'inscription
+            $inscri = $eleves->isValidationInscription();
+            $inscription = $form->get('validation_inscription')->getData();
+
+            if($inscription == true && $inscri == false){
+                $fuseauHoraire = new DateTimeZone('Europe/Paris');
+                $dateActuelle = new DateTime('now', $fuseauHoraire);
+                $eleves->setValidationInscription(true);
+            }
+
+
+
+            $entityManager->persist($eleves);
+            $entityManager->flush();
+        }
+
+
+        //dn = date de naissance
+        $dn = $eleves->getDateNaissance();
+        if(!empty($dn)){
+            $dn = $dn->format('d/m/Y');
+        }
+
+        //di = date d'inscription
+        $inscri = $eleves->isValidationInscription();
+        $di = $eleves->getDateInscription();
+        if(!empty($di) && $inscri == true){
+            $di = $di->format('d/m/Y');
+        } else {
+            $di = '';
+        }
+
+        return $this->render('main/eleve.html.twig', [
             'elevesRepository' => $elevesRepository->findBy([], ['nom' => 'ASC']),
-            'eleves' => $eleves
+            'elevesForm' => $form->createView(),
         ]);
     }
 
@@ -45,8 +134,11 @@ class MainController extends AbstractController
                 // on récupère l'ancienne photo
                 $previmage = $eleves->getPhoto();
 
-                // on supprime l'ancienne photo de profil
-                $pictureService->delete($previmage, $folder);
+                // on supprime l'ancienne photo de profil si elle existe
+                if(!empty($previmage)){
+                    $pictureService->delete($previmage, $folder);
+                }
+
 
                 // on ajoute la nouvelle photo de profil
                 $fichier = $pictureService->add($image, $folder, 300, 300);
